Make the json loader and parser independent from the generated ir files.
