+++ ir/node.h
@@ -19,11 +19,10 @@ limitations under the License.
 
 #include <iosfwd>
 
-#include "ir-tree-macros.h"
 #include "ir/gen-tree-macro.h"
-#include "lib/castable.h"
+#include "ir/inode.h"
+#include "ir/ir-tree-macros.h"
 #include "lib/cstring.h"
-#include "lib/exceptions.h"
 #include "lib/source_file.h"
 
 namespace P4 {
@@ -50,46 +49,6 @@ template <class T>
 class Vector;  // IWYU pragma: keep
 template <class T>
 class IndexedVector;  // IWYU pragma: keep
-
-/// SFINAE helper to check if given class has a `static_type_name`
-/// method. Definite node classes have them while interfaces do not
-template <class, class = void>
-struct has_static_type_name : std::false_type {};
-
-template <class T>
-struct has_static_type_name<T, std::void_t<decltype(T::static_type_name())>> : std::true_type {};
-
-template <class T>
-inline constexpr bool has_static_type_name_v = has_static_type_name<T>::value;
-
-// node interface
-class INode : public Util::IHasSourceInfo, public IHasDbPrint, public ICastable {
- public:
-    virtual ~INode() {}
-    virtual const Node *getNode() const = 0;
-    virtual Node *getNode() = 0;
-    virtual void toJSON(JSONGenerator &) const = 0;
-    virtual cstring node_type_name() const = 0;
-    virtual void validate() const {}
-
-    // default checkedTo implementation for nodes: just fallback to generic ICastable method
-    template <typename T>
-    std::enable_if_t<!has_static_type_name_v<T>, const T *> checkedTo() const {
-        return ICastable::checkedTo<T>();
-    }
-
-    // alternative checkedTo implementation that produces slightly better error message
-    // due to node_type_name() / static_type_name() being available
-    template <typename T>
-    std::enable_if_t<has_static_type_name_v<T>, const T *> checkedTo() const {
-        const auto *result = to<T>();
-        BUG_CHECK(result, "Cast failed: %1% with type %2% is not a %3%.", this, node_type_name(),
-                  T::static_type_name());
-        return result;
-    }
-
-    DECLARE_TYPEINFO_WITH_TYPEID(INode, NodeKind::INode);
-};
 
 class Node : public virtual INode {
  public:
