+++ ir/visitor.cpp
@@ -23,7 +23,6 @@ limitations under the License.
 #include "absl/container/flat_hash_map.h"
 #include "absl/time/clock.h"
 #include "absl/time/time.h"
-#include "ir/ir-generated.h"
 #include "lib/hash.h"
 
 #if HAVE_CXXABI_H
